Understanding Systematic Reviews and Meta-Analyses
Systematic reviews should be distinguished from narrative reviews. In the latter, an editor asks an expert to sum up all of the information 
that is known about a particular topic. However, the expert is under no constraints regarding what he or she does, or does not, choose to 
include in the review. As a result, his or her bias can influence the final message. A systematic review, which may or may not be written by 
experts, typically asks a narrower question, and then answers it using the entirety of the medical literature. The systematic review process 
includes computer searches to identify the pertinent literature, a statement of the inclusion and exclusion criteria for identified studies, a 
list of items of interest to extract from each study, a method to assess the quality of each study, a summary of the evidence that has been 
found (which may or may not involve attempts to combine data), a discussion of the evidence and the limitations of the conclusions, and 
suggestions for future research efforts. If the data are combined, that process is called meta-analysis. In meta-analysis, an estimate of the 
reliability of each study is made, and those that appear to be more reliable are weighed more heavily when the data are combined. While 
systematic reviews depend on a more preplanned method and thus, unlike narrative reviews, contain sections on method, they can be 
easily read once the reader becomes familiar with the vocabulary